collect reduction tests
